+++ src/entities/userPayments/UserPayments.tsx
@@ -1,9 +1,6 @@
 "use client"
 import s from './UserPayemnts.module.scss'
 import {useQueryParams} from "@/shared/lib/hooks/useQueryParams";
-import {useGetOneUserQuery} from "@/graphql/queries/getOneUserData.generated";
-import {useGetPostsByUserQuery} from "@/graphql/queries/getPostsByUser.generated";
-import {useGetAllPostsQuery} from "@/graphql/queries/getAllPosts.generated";
 import {SortDirection} from "@/graphql/types";
 import {useGetAllPaymentsQuery} from "@/graphql/queries/payments/getAllPayments.generated";
 import {useGetPaymentsByUserQuery} from "@/graphql/queries/payments/getPaymentsByUser.generated";
