+++ src/graphql/queries/posts/getAllPosts.ts
