+++ src/graphql/queries/userData/getOneUserData.ts
